+++ vrp-pragmatic/tests/regression/break_test.rs
@@ -116,7 +116,7 @@ fn can_handle_properly_invalid_break_removal() {
                     type_id: "vehicle1".to_string(),
                     vehicle_ids: vec!["vehicle1_1".to_string()],
                     profile: VehicleProfile { matrix: "car".to_string(), scale: None },
-                    costs: VehicleCosts { fixed: Some(20.), distance: TieredCost::Fixed(0.002), time: TieredCost::Fixed(0.003) },
+                    costs: VehicleCosts { fixed: Some(20.), distance: TieredCost::Fixed(0.002), time: TieredCost::Fixed(0.003), calculation_mode: None },
                     shifts: vec![VehicleShift {
                         start: ShiftStart {
                             earliest: "2020-07-04T09:00:00Z".to_string(),