+++ vrp-pragmatic/tests/regression/pd_same_stops.rs
@@ -36,7 +36,7 @@ fn can_handle_two_pd_jobs_with_same_locations_and_unusual_routing() {
         },
         fleet: Fleet {
             vehicles: vec![VehicleType {
-                costs: VehicleCosts { fixed: None, distance: TieredCost::Fixed(0.0), time: TieredCost::Fixed(1.0) },
+                costs: VehicleCosts { fixed: None, distance: TieredCost::Fixed(0.0), time: TieredCost::Fixed(1.0), calculation_mode: None },
                 shifts: vec![VehicleShift {
                     start: ShiftStart {
                         earliest: format_time(0.),